+++ assets/js/router.js
@@ -1,4 +1,5 @@
 const routes = {
+  "systemd-builder": () => import("../../modules/linux/systemd-builder/index.js"),
   "cron-generator": () => import("../../modules/linux/cron-generator/index.js"),
   "chmod-calculator": () => import("../../modules/linux/chmod-calculator/index.js"),
   "url-encoder": () => import("../../modules/developer/url-encoder/index.js"),